+++ src/managers/dnf.rs
@@ -1,6 +1,6 @@
 use std::{fmt::Display, process::Command};
 
-use crate::{Cmd, Commands, Package, PackageManager, RepoError};
+use crate::{Cmd, Commands, Package, PackageManager};
 
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
@@ -33,20 +33,16 @@ impl PackageManager for DandifiedYUM {
         }
     }
 
-    fn add_repo(&self, repo: &str) -> Result<(), RepoError> {
-        if !self
-            .install(Package::new("dnf-command(config-manager)", None))
-            .success()
-        {
-            return Err(RepoError::with_msg(
-                "failed to install config-manager plugin",
-            ));
-        }
+    fn add_repo(&self, repo: &str) -> anyhow::Result<()> {
+        anyhow::ensure!(
+            self.install(Package::new("dnf-command(config-manager)", None))
+                .success(),
+            "failed to install config-manager plugin"
+        );
 
-        self.exec_cmds_status(&self.consolidated(Cmd::AddRepo, &[repo]))
-            .success()
-            .then_some(())
-            .ok_or(RepoError::default())
+        let s = self.exec_cmds_status(&self.consolidated(Cmd::AddRepo, &[repo]));
+        anyhow::ensure!(s.success(), "failed to add repo");
+        Ok(())
     }
 }
