+++ src/managers/yum.rs
@@ -35,7 +35,7 @@ impl PackageManager for YellowdogUpdaterModified {
     fn parse_pkg<'a>(&self, line: &str) -> Option<crate::Package> {
         self.0.parse_pkg(line)
     }
-    fn add_repo(&self, repo: &str) -> Result<(), crate::RepoError> {
+    fn add_repo(&self, repo: &str) -> anyhow::Result<()> {
         self.0.add_repo(repo)
     }
 }
